--- homeweather/src/main/kotlin/com/babkovic/current/config/CurrentWeatherJacksonConfig.kt
@@ -1,28 +1,0 @@
-package com.babkovic.current.config
-
-import com.babkovic.current.mapper.CurrentWeatherDeserializer
-import com.babkovic.current.mapper.CurrentWeatherSerializer
-import com.babkovic.current.model.domain.CurrentWeather
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.module.SimpleModule
-import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration(proxyBeanMethods = false)
-class CurrentWeatherJacksonConfig {
-    @Bean
-    fun openWeatherObjectMapper(@Qualifier("objectMapper") initialObjectMapper: ObjectMapper): ObjectMapper {
-        val objectMapper = ObjectMapper()
-        registerCurrentWeather(objectMapper)
-
-        return objectMapper
-    }
-
-    private fun registerCurrentWeather(objectMapper: ObjectMapper) {
-        val module = SimpleModule()
-        module.addDeserializer(CurrentWeather::class.java, CurrentWeatherDeserializer())
-        module.addSerializer(CurrentWeather::class.java, CurrentWeatherSerializer())
-        objectMapper.registerModule(module)
-    }
-}
