--- homeweather/src/main/kotlin/com/babkovic/current/exception/CurrentWeatherException.kt
@@ -1,6 +1,0 @@
-package com.babkovic.current.exception
-
-class CurrentWeatherException : RuntimeException {
-    constructor(message: String?, throwable: Throwable) : super(message, throwable)
-    constructor(message: String?) : super(message)
-}
