--- homeweather/src/main/kotlin/com/babkovic/current/service/CurrentWeatherService.kt
@@ -1,6 +1,0 @@
-package com.babkovic.current.service
-
-import org.springframework.stereotype.Service
-
-@Service
-annotation class CurrentWeatherService
