+++ homeweather/src/main/kotlin/com/babkovic/home/Application.kt
@@ -1,23 +1,13 @@
-package com.babkovic.smssender;
+package com.babkovic.home
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories
 
-@SpringBootApplication(
-    scanBasePackages = [
-        "com.babkovic.smssender",
-        "com.babkovic.config",
-        "com.babkovic.home.current.model.domain",
-    ]
-)
-@EnableReactiveMongoRepositories(
-    basePackages = [
-        "com.babkovic.home.current.model.repository",
-        "com.babkovic.smssender.model.repository"
-    ]
-)
-
+@SpringBootApplication(scanBasePackages = ["com.babkovic.home.current", "com.babkovic.config"])
+@EnableReactiveMongoRepositories(basePackages = ["com.babkovic.home.current.model.repository"])
+@ConfigurationPropertiesScan("com.babkovic.config")
 class Application {
 
     companion object {
