+++ homeweather/src/main/kotlin/com/babkovic/home/current/controller/CurrentWeatherControllerImpl.kt
@@ -1,0 +1,19 @@
+package com.babkovic.home.current.controller
+
+import com.babkovic.home.current.model.domain.CurrentWeather
+import com.babkovic.home.current.service.ICurrentWeatherService
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.web.bind.annotation.RestController
+import reactor.core.publisher.Flux
+import reactor.core.publisher.Mono
+
+@RestController
+class CurrentWeatherControllerImpl(@Autowired private val service: ICurrentWeatherService) : CurrentWeatherController {
+    override fun saveCurrentWeather(): Flux<CurrentWeather> {
+        return service.saveCurrentWeather()
+    }
+
+    override fun saveCurrentWeather(lat: Double, lon: Double): Mono<CurrentWeather> {
+        return service.saveCurrentWeather(lat, lon)
+    }
+}
