+++ homeweather/src/main/kotlin/com/babkovic/home/current/exception/CurrentWeatherException.kt
@@ -1,0 +1,6 @@
+package com.babkovic.home.current.exception
+
+class CurrentWeatherException : RuntimeException {
+    constructor(message: String?, throwable: Throwable) : super(message, throwable)
+    constructor(message: String?) : super(message)
+}
