+++ homeweather/src/main/kotlin/com/babkovic/home/current/exception/handler/CurrentWeatherApiExceptionHandler.kt
@@ -1,0 +1,29 @@
+package com.babkovic.home.current.exception.handler
+
+import com.babkovic.home.current.controller.CurrentWeatherControllerImpl
+import com.babkovic.home.current.exception.CurrentWeatherException
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.reactive.result.method.annotation.ResponseEntityExceptionHandler
+import org.springframework.web.server.ServerWebExchange
+import reactor.core.publisher.Mono
+
+@ControllerAdvice(basePackageClasses = [CurrentWeatherControllerImpl::class])
+class CurrentWeatherApiExceptionHandler : ResponseEntityExceptionHandler() {
+    @ResponseBody
+    @ExceptionHandler(CurrentWeatherException::class)
+    fun handleOpenWeatherApiException(
+        request: ServerWebExchange, ex: CurrentWeatherException
+    ): Mono<*> {
+        return super.handleExceptionInternal(
+            ex,
+            ex.message,
+            HttpHeaders(),
+            HttpStatus.BAD_REQUEST,
+            request
+        )
+    }
+}
