+++ homeweather/src/main/kotlin/com/babkovic/home/current/service/CurrentWeatherService.kt
@@ -1,0 +1,6 @@
+package com.babkovic.home.current.service
+
+import org.springframework.stereotype.Service
+
+@Service
+annotation class CurrentWeatherService
