+++ homeweather/src/test/kotlin/com/babkovic/home/current/controller/CurrentWeatherIT.kt
@@ -1,33 +1,48 @@
-package com.babkovic.openweather.controller
+package com.babkovic.home.current.controller
 
-import com.babkovic.openweather.BaseTest
+import com.babkovic.home.BaseTest
 import com.babkovic.home.current.model.domain.CurrentWeather
+import com.babkovic.home.current.model.repository.CurrentWeatherRepository
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInfo
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.MediaType
-import org.springframework.test.web.reactive.server.FluxExchangeResult
 import reactor.kotlin.test.test
-import reactor.test.StepVerifier
 import kotlin.test.assertEquals
 
-class OpenWeatherControllerIT : BaseTest() {
+class CurrentWeatherIT : BaseTest() {
 
     companion object {
-        private val LOGGER: Logger = LoggerFactory.getLogger(OpenWeatherControllerIT::class.java)
+        private val LOGGER: Logger = LoggerFactory.getLogger(CurrentWeatherIT::class.java)
     }
 
+    @AfterEach
+    fun cleanUp() {
+        repo.deleteAll().block()
+    }
+
+    @Autowired
+    private lateinit var repo: CurrentWeatherRepository
+
     @Test
-    fun `should return current, when calling get current weather from open weather by coords`(testInfo: TestInfo) {
+    fun `should return and persist current weather mono when calling save endpoint`(testInfo: TestInfo) {
         LOGGER.info("Starting test ${testInfo.displayName}\n")
 
         //given and when
-        val url = "$GATEWAY_URL/open/current?lat=49.136372&lon=20.24386"
-        val resultFlux: FluxExchangeResult<CurrentWeather> = callCurrentWeatherFlux(url)
+        val resFlux = client.post()
+            .uri("${GATEWAY_URL}/home/current/save?lat=49.136372&lon=20.24386")
+            .accept(MediaType.APPLICATION_NDJSON)
+            .exchange()
+            .expectStatus().isOk
+            .expectHeader().contentType(MediaType.APPLICATION_NDJSON)
+            .returnResult(CurrentWeather::class.java)
 
         //then
-        StepVerifier.create(resultFlux.responseBody)
+        resFlux.responseBody.test()
             .assertNext { currentWeather ->
                 run {
                     assertEquals(20.2439, currentWeather.lat, "latitude is not the same")
@@ -35,21 +50,26 @@ class OpenWeatherControllerIT : BaseTest() {
                     assertEquals("Asia/Riyadh", currentWeather.timezone, "timezone is not the same")
                 }
             }
-
             .thenCancel()
             .verify()
+        Assertions.assertEquals(1, repo.findAll().count().block())
 
         LOGGER.info("Ending test ${testInfo.displayName}\n")
 
     }
 
     @Test
-    fun `should return current flux, when calling get current weather from open weather`(testInfo: TestInfo) {
+    fun `should save and remove entity from db when calling current weather by coords`(testInfo: TestInfo) {
         LOGGER.info("Starting test ${testInfo.displayName}\n")
 
         //given and when
-        val url = "$GATEWAY_URL/open/current/all"
-        val resFlux: FluxExchangeResult<CurrentWeather> = callCurrentWeatherFlux(url)
+        val resFlux = client.post()
+            .uri("${GATEWAY_URL}/home/current/bulk/save")
+            .accept(MediaType.APPLICATION_NDJSON)
+            .exchange()
+            .expectStatus().isOk
+            .expectHeader().contentType(MediaType.APPLICATION_NDJSON)
+            .returnResult(CurrentWeather::class.java)
 
         //then
         resFlux.responseBody.test()
@@ -69,16 +89,9 @@ class OpenWeatherControllerIT : BaseTest() {
             }
             .thenCancel()
             .verify()
+        Assertions.assertEquals(2, repo.findAll().count().block())
 
         LOGGER.info("Ending test ${testInfo.displayName}\n")
 
     }
-
-    private fun callCurrentWeatherFlux(urlToCall: String) = client.get()
-        .uri(urlToCall)
-        .accept(MediaType.APPLICATION_NDJSON)
-        .exchange()
-        .expectStatus().isOk
-        .expectHeader().contentType(MediaType.APPLICATION_NDJSON)
-        .returnResult(CurrentWeather::class.java)
 }
