+++ openweather/src/main/kotlin/com/babkovic/openweather/service/IOpenWeatherService.kt
@@ -1,6 +1,6 @@
 package com.babkovic.openweather.service
 
-import com.babkovic.current.model.domain.CurrentWeather
+import com.babkovic.home.current.model.domain.CurrentWeather
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
