+++ smssender/src/main/kotlin/com/babkovic/smssender/model/domain/Coord.kt
@@ -1,0 +1,6 @@
+package com.babkovic.smssender.model.domain
+
+data class Coord(
+    val lon: Double,
+    val lat: Double
+)
