+++ smssender/src/main/kotlin/com/babkovic/smssender/service/ITwilioService.kt
@@ -1,6 +1,6 @@
 package com.babkovic.smssender.service
 
-import com.babkovic.current.model.domain.CurrentWeather
+import com.babkovic.home.current.model.domain.CurrentWeather
 import com.babkovic.smssender.model.domain.User
 import com.twilio.rest.api.v2010.account.Message
 import reactor.core.publisher.Mono
