+++ smssender/src/test/resources/application-test.yml
@@ -8,5 +8,8 @@ twilio:
 messenger:
   schema: "http"
   sub-domain: "localhost"
+weather:
+  open-weather-api-url: "https://api.openweathermap.org/data/3.0/onecall"
+  weather-plan-open-url: "http://localhost:8083/open/current"
 gateway:
   port: "8083"
